Updated to version 2.0.0

+++ mute/__init__.py
@@ -1,4 +1,7 @@
 import os
+from shutil import copytree, rmtree
+
+GitHub_data_file = "data_20220715"
 
 
 def download_file(url, dir_out):
@@ -25,22 +28,36 @@ def download_file(url, dir_out):
 
         f.extractall(dir_out)
 
+    # Rename the data directory "data"
+    # If it already exists, move the unzipped files into it and replace existing ones with the same name
+    # If it does not, rename the unzipped directory "data"
+
     os.chdir(dir_out)
-    os.rename("data_20211219", "data")
+
+    if os.path.isdir("data"):
+
+        copytree(GitHub_data_file, "data", dirs_exist_ok=True)
+        rmtree(GitHub_data_file)
+
+    else:
+
+        os.rename(GitHub_data_file, "data")
+
     os.chdir("..")
 
     # Delete the zip file
 
-    os.remove(file_out)
+    if os.path.isfile(file_out):
+        os.remove(file_out)
 
 
-# Download the files from GitHub
+# Download the files from GitHub to the directory MUTE is being run from
 
 if not os.path.isfile(
-    os.path.join(os.path.dirname(__file__), "data", "data_20211219.txt")
+    os.path.join(os.path.dirname(__file__), "data", f"{GitHub_data_file}.txt")
 ):
 
     download_file(
-        "https://github.com/wjwoodley/mute/releases/download/0.1.0/data_20211219.zip",
+        f"https://github.com/wjwoodley/mute/releases/download/0.1.0/{GitHub_data_file}.zip",
         os.path.dirname(__file__),
     )
